x86: mm: Skip faulting instruction for VM_DROPPABLE faults
Currently, VM_DROPPABLE allocations fault in a limited way where the
faulting instruction was retried instead of skipped. Finish that up with
the platform-specific aspect of skipping the actual instruction.

This works on x86, where instructions are variable size, by copying
userspace's %rip to a stack buffer of size MAX_INSN_SIZE, decoding it,
and then adding the length of the decoded instruction to userspace's
%rip. In the event any of these fail, just fallback to not advancing
%rip and trying again.

Cc: linux-mm@kvack.org
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>