mm: Do not OOM when failing VM_DROPPABLE faults
In addition to the requirements mentioned in "mm: add VM_DROPPABLE for
designating always lazily freeable mappings", it is also desirable to
have one additional property:

- It shouldn't reserve actual memory, but it also shouldn't crash when
  page faulting in memory if none is available
  * Uh-oh: MAP_NORESERVE respects vm.overcommit_memory=2.
  * Uh-oh: VM_NORESERVE means segfaults.

So, make the allocations not warn or retry, and don't OOM when they
fail. This in turn will cause userspace to just keep retrying the
faulting instruction until it works, and later on we can have it skip
the faulting instruction instead.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>