mm: re-add pte_dirty() with vma argument
The implementation of pte_dirty() currently forwards the pte argument to
pte_dirty_novma() and discards the vma argument, but this now allows the
vma to be used in future implementations of pte_dirty(). Convert
relevant callers who use pte_dirty() in a context where a vma is
available.

This change functionally does nothing.

Cc: linux-mm@kvack.org
Suggested-by: Linus Torvalds <torvalds@linux-foundation.org>
Suggested-by: David Hildenbrand <dhildenb@redhat.com>
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>